Refactor: Consolidate Progress and Controls in one VC
Extraxcting the ArtworkControls along the way. This turns out to be a better abstraction after all.

Remove unused file
